requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.10.0
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.5
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
clr==1.0.3
colorama==0.4.6
comtypes==1.4.11
confection==0.1.5
contourpy==1.3.3
coqpit==0.0.17
cutlet==0.5.0
cycler==0.12.1
cymem==2.0.11
Cython==3.1.3
dateparser==1.1.8
decorator==5.2.1
distro==1.9.0
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
fastapi==0.115.14
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
fugashi==1.5.1
g2pkk==0.1.2
grpcio==1.74.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
inflect==7.5.0
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonlines==1.2.0
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
marisa-trie==1.3.0
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mojimoji==0.0.13
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
murmurhash==1.0.13
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.61.2
numpy==1.26.4
openai==1.101.0
packaging==25.0
pandas==1.5.3
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.55.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
PyYAML==6.0.2
rank-bm25==0.2.2
regex==2025.7.34
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
shellingham==1.5.4
simpleaudio==1.0.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tqdm==4.67.1
trainer==0.0.36
transformers==4.41.1
TTS==0.22.0
typeguard==4.4.4
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.9.post2
Unidecode==1.4.0
unidic-lite==1.0.8
urllib3==2.5.0
uvicorn==0.30.6
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
